+++ test/cli.js
@@ -5,7 +5,7 @@ process.chdir('../');
 
 test('grunion - pass', async t => {
   const result = await execa(
-    './bin/grunion',
+    './src/cmd.js',
     ['./test/fixtures/*-pass.js'],
     {preferLocal: true}
   );
@@ -17,7 +17,7 @@ test('grunion - pass', async t => {
 
 test('grunion - fail', async t => {
   try {
-    await execa('./bin/grunion',
+    await execa('./src/cmd.js',
      ['./test/fixtures/*.js'],
      {preferLocal: true}
    );
@@ -33,7 +33,7 @@ test('grunion - fail', async t => {
 test('grunion - fail - fail-fast', async t => {
   try {
     await execa(
-      './bin/grunion',
+      './src/cmd.js',
       ['./test/fixtures/*.js', '--fail-fast'],
       {preferLocal: true}
     );
@@ -45,7 +45,7 @@ test('grunion - fail - fail-fast', async t => {
 
 test('grunion - pass - babel-node', async t => {
   const result = await execa(
-    './bin/grunion',
+    './src/cmd.js',
     ['./test/fixtures/*-pass.js', '--run', 'babel-node <%= file.path %>'],
     {preferLocal: true}
   );
@@ -58,7 +58,7 @@ test('grunion - pass - babel-node', async t => {
 test('grunion - fail - babel-node', async t => {
   try {  // returns error
     await execa(
-      './bin/grunion',
+      './src/cmd.js',
       ['./test/fixtures/*.js', '--run', 'babel-node <%= file.path %>'],
       {preferLocal: true}
     );
@@ -74,7 +74,7 @@ test('grunion - fail - babel-node', async t => {
 test('grunion - fail - fail-fast babel-node', async t => {
   try {
     await execa(
-      './bin/grunion',
+      './src/cmd.js',
       ['./test/fixtures/*.js', '--run', 'babel-node <%= file.path %>', '--fail-fast'],
       {preferLocal: true}
     );
@@ -86,7 +86,7 @@ test('grunion - fail - fail-fast babel-node', async t => {
 
 test('grunion - echo', async t => {
   const result = await execa(
-    './bin/grunion',
+    './src/cmd.js',
     ['./test/fixtures/*.js', '--run', 'echo <%= file.path %>'],
     {preferLocal: true}
   );
@@ -98,9 +98,12 @@ test('grunion - echo', async t => {
 
 test('grunion - echo - serial', async t => {
   const result = await execa(
-    './bin/grunion',
+    './src/cmd.js',
     ['./test/fixtures/*.js', '--run', 'echo <%= file.path %>', '--serial'],
     {preferLocal: true}
   );
-  t.is(result.stdout, './test/fixtures/a-pass.js\n./test/fixtures/b-pass.js\n./test/fixtures/c-fail.js\n./test/fixtures/d.js');
+  t.regex(result.stdout, /.\/test\/fixtures\/a-pass\.js/);
+  t.regex(result.stdout, /.\/test\/fixtures\/b-pass\.js/);
+  t.regex(result.stdout, /.\/test\/fixtures\/c-fail\.js/);
+  t.regex(result.stdout, /.\/test\/fixtures\/d\.js/);
 });